+++ lib/turbolinks-animate/engine.rb
@@ -1,0 +1,6 @@
+require 'rails/railtie'
+
+module TurbolinksAnimate
+    class Engine < ::Rails::Engine
+    end
+end