+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 * nothing yet
 
+### 1.2.0 - 2018-01-06
+
+* enhancements
+    * automatically include `TurbolinksAnimateHelper` in controllers
+    * updated [turbolinks-animate.js](https://github.com/jonhue/turbolinks-animate.js) to `1.3.8`
+
 ### 1.1.0 - 2017-12-17
 
 * enhancements